icu_discharge/requirements.txt
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
catboost==1.2.8
optuna==4.3.0
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0